+++ src/components/movies_section/MovieCard.css
