+++ src/components/movies_section/MovieList.css
