+++ src/utils/tmdbOptions.js
@@ -1,0 +1,9 @@
+const getTMDBOptions = () => ({
+  method: "GET",
+  headers: {
+    accept: "application/json",
+    Authorization: `Bearer ${import.meta.env.VITE_API_KEY}`,
+  },
+});
+
+export default getTMDBOptions;
